FY-31/feat: logout view, refactor register, login view

+++ authentication/views.py
@@ -4,7 +4,6 @@ from django.views import View
 from django.contrib.auth import login, logout
 from rest_framework.authtoken.models import Token
 from django.contrib.auth import authenticate
-
 from accounts.models import User
 from accounts.serializers import UserSerializer
 
@@ -18,11 +17,44 @@ from django.db import IntegrityError
 import json
 
 
+async def load_json_data(request):
+    try:
+        return json.loads(request.body), None
+    except json.decoder.JSONDecodeError:
+        return None, JsonResponse(
+            {"status": False, "error": "Invalid JSON"}, status=400
+        )
+
+
+@database_sync_to_async
+def authenticate_user_and_token(intra_id, request):
+    try:
+        token_key = (
+            request.META.get("HTTP_AUTHORIZATION", "").split()[1]
+            if "HTTP_AUTHORIZATION" in request.META
+            else None
+        )
+        user = User.objects.get(intra_id=intra_id)
+        token = Token.objects.get(user=user, key=token_key)
+        return (
+            user,
+            None if token else None,
+            JsonResponse({"status": False, "error": "Invalid Token"}, status=401),
+        )
+    except (User.DoesNotExist, Token.DoesNotExist):
+        return (
+            None,
+            JsonResponse({"status": False, "error": "Invalid Token"}, status=401),
+        )
+
+
 @method_decorator(csrf_exempt, name="dispatch")
 class RegisterView(View):
     async def post(self, request, *args, **kwargs):
-        serializer = UserSerializer(data=json.loads(request.body))
-
+        data, error = await load_json_data(request)
+        if error:
+            return error
+        serializer = UserSerializer(data=data)
         if await sync_to_async(serializer.is_valid)():
             user = await self.create_user(serializer.validated_data)
             if user:
@@ -44,36 +76,40 @@ class RegisterView(View):
 
     async def get(self, request, *args, **kwargs):
         users = await self.get_users()
-        users_dict = await sync_to_async(model_to_dict)(users)
-        return JsonResponse({"status": True, "data": users_dict}, status=200)
+        return JsonResponse({"status": True, "data": users}, status=200)
 
     @database_sync_to_async
     def get_users(self):
-        users = User.objects.all()
-        return [user.intra_id for user in users]
+        return list(User.objects.all().values("intra_id", "email"))
 
     async def delete(self, request, *args, **kwargs):
-        intra_id = json.loads(request.body)["intra_id"]
-        success = await self.delete_user(intra_id)
+        data, error = await load_json_data(request)
+        if error:
+            return error
+        intra_id = data.get("intra_id")
+        user, error = await authenticate_user_and_token(intra_id, request)
+        if error:
+            return error
+        success = await self.delete_user(user)
         if success:
             return JsonResponse({"status": True}, status=200)
         else:
             return JsonResponse({"status": False}, status=400)
 
     @database_sync_to_async
-    def delete_user(self, intra_id):
-        try:
-            user = User.objects.get(intra_id=intra_id)
+    def delete_user(self, user):
+        if user:
             user.delete()
             return True
-        except User.DoesNotExist:
-            return False
+        return False
 
 
 @method_decorator(csrf_exempt, name="dispatch")
 class LoginView(View):
     async def post(self, request, *args, **kwargs):
-        data = json.loads(request.body)
+        data, error = await load_json_data(request)
+        if error:
+            return error
         user = await self.authenticate_user(data["intra_id"], data["password"])
         if user:
             await self.login_user(request, user)
@@ -100,16 +136,47 @@ class LoginView(View):
         login(request, user)
 
     async def get(self, request, *args, **kwargs):
-        user = await self.get_user(request)
-        user_dict = await sync_to_async(model_to_dict)(user)
-        user_dict.pop("password", None)
+        data, error = await load_json_data(request)
+        if error:
+            return error
+        user = await self.get_user(data.get("intra_id"), request)
         if user:
+            user_dict = await sync_to_async(model_to_dict)(user)
+            user_dict.pop("password", None)
             return JsonResponse({"status": True, "data": user_dict}, status=200)
         else:
-            return JsonResponse({"status": False}, status=400)
+            return JsonResponse({"status": False}, status=401)
 
     @database_sync_to_async
-    def get_user(self, request):
-        if request.user.is_authenticated:
-            return request.user
-        return None
+    def get_user(self, intra_id, request):
+        try:
+            user = User.objects.get(intra_id=intra_id)
+            logged_in_user_id = request.session.get("_auth_user_id")
+            return user if str(user.intra_id) == logged_in_user_id else None
+        except User.DoesNotExist:
+            return None
+
+
+@method_decorator(csrf_exempt, name="dispatch")
+class LogoutView(View):
+    async def post(self, request, *args, **kwargs):
+        data, error = await load_json_data(request)
+        if error:
+            return error
+        user, error = await authenticate_user_and_token(data["intra_id"], request)
+        if error:
+            return error
+        logged_out = await self.logout_user(user, request)
+        if logged_out:
+            return JsonResponse({"status": True}, status=200)
+        return JsonResponse({"status": False, "error": "user not found"}, status=401)
+
+    @database_sync_to_async
+    def logout_user(self, user, request):
+        try:
+            if user == request.user:
+                logout(request)
+                return True
+            return False
+        except User.DoesNotExist:
+            return False
